+++ src/open_stocks_mcp/tools/robinhood_market_data_tools.py
@@ -21,6 +21,7 @@ from open_stocks_mcp.tools.market.movers import (
     get_top_movers,
     get_top_movers_sp500,
 )
+from open_stocks_mcp.tools.market.news import get_stock_news
 from open_stocks_mcp.tools.market.ratings import get_stock_ratings
 from open_stocks_mcp.tools.rate_limiter import get_rate_limiter
 
@@ -109,73 +110,6 @@ async def get_stock_earnings(symbol: str) -> dict[str, Any]:
     except Exception as e:
         logger.error(f"Failed to get earnings for {symbol}: {e}")
         return create_error_response(e, "get_stock_earnings")
-
-
-@handle_robin_stocks_errors
-async def get_stock_news(symbol: str) -> dict[str, Any]:
-    """Get news stories for a stock.
-
-    Args:
-        symbol: Stock ticker symbol (e.g., "AAPL")
-
-    Returns:
-        JSON object with news stories in "result" field:
-        {
-            "result": {
-                "symbol": "AAPL",
-                "news": [
-                    {
-                        "title": "Apple Reports Strong Q2 Results",
-                        "author": "Tech News Reporter",
-                        "published_at": "2024-07-09T14:30:00Z",
-                        "source": "TechCrunch",
-                        "summary": "Apple exceeded expectations...",
-                        "url": "https://...",
-                        "preview_image_url": "https://...",
-                        "num_clicks": 1250
-                    }
-                ],
-                "count": 20
-            }
-        }
-    """
-    try:
-        # Input validation
-        if not validate_symbol(symbol):
-            return create_error_response(
-                ValueError(f"Invalid symbol format: {symbol}"), "symbol validation"
-            )
-
-        symbol = symbol.strip().upper()
-
-        # Ensure authenticated
-        session_mgr = get_session_manager()
-        if not await session_mgr.ensure_authenticated():
-            return create_error_response(
-                ValueError("Authentication required"), "authentication"
-            )
-
-        # Apply rate limiting
-        rate_limiter = get_rate_limiter()
-        await rate_limiter.acquire()
-
-        log_api_call("get_stock_news", symbol=symbol)
-
-        # Get news with retry logic
-        news_data = await execute_with_retry(rh.get_news, symbol)
-
-        if not news_data:
-            return create_no_data_response(
-                f"No news data found for symbol: {symbol}", {"symbol": symbol}
-            )
-
-        return create_success_response(
-            {"symbol": symbol, "news": news_data, "count": len(news_data)}
-        )
-
-    except Exception as e:
-        logger.error(f"Failed to get news for {symbol}: {e}")
-        return create_error_response(e, "get_stock_news")
 
 
 @handle_robin_stocks_errors
