+++ tests/unit/test_market_research_tools.py
@@ -6,12 +6,12 @@ from unittest.mock import AsyncMock, patch
 import pytest
 
 from open_stocks_mcp.tools.market.movers import get_top_movers_sp500
+from open_stocks_mcp.tools.market.news import get_stock_news
 from open_stocks_mcp.tools.market.ratings import get_stock_ratings
 from open_stocks_mcp.tools.robinhood_market_data_tools import (
     get_stock_earnings,
     get_stock_events,
     get_stock_level2_data,
-    get_stock_news,
     get_stock_splits,
 )
 
@@ -399,9 +399,9 @@ class TestStockEarnings:
 class TestStockNews:
     """Test stock news functionality."""
 
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.news.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.news.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.news.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
@@ -470,9 +470,9 @@ class TestStockNews:
 
     @pytest.mark.exception_test
     @pytest.mark.skip(reason="Slow exception test - run with pytest -m exception_test")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.execute_with_retry")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_session_manager")
-    @patch("open_stocks_mcp.tools.robinhood_market_data_tools.get_rate_limiter")
+    @patch("open_stocks_mcp.tools.market.news.execute_with_retry")
+    @patch("open_stocks_mcp.tools.market.news.get_session_manager")
+    @patch("open_stocks_mcp.tools.market.news.get_rate_limiter")
     @pytest.mark.journey_research
     @pytest.mark.unit
     @pytest.mark.asyncio
